+++ .github/workflows/publish-npm.yml
@@ -31,15 +31,12 @@ jobs:
           registry-url: https://registry.npmjs.com/
       - name: Get version
         run: echo "TAG_VERSION=${GITHUB_REF#refs/*/}" >> $GITHUB_ENV
-      - name: build
-        run: yarn fast-build
-        env:
-          TAG_VERSION: ${{env.TAG_VERSION}}
-      - name: generate npm token
+        run: echo "GIT_HEAD=${GITHUB_SHA}" >> GITHUB_ENV
+      - name: build&publish
         run: echo "//registry.npmjs.com/:_authToken=${{ secrets.NPM_PUBLISH_TOKEN }}" > ./.npmrc
-      - name: Publish npm
-        run: sh ./scripts/publish.sh
+        run: sh ./scripts/monorepo.sh
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_PUBLISH_TOKEN}}
           TAG_VERSION: ${{env.TAG_VERSION}}
+          GIT_HEAD: ${{env.GIT_HEAD}}
           REGISTRY: https://registry.npmjs.com/
