+++ package.json
@@ -14,6 +14,7 @@
     "build:locale-umd": "node ./build/build-locale.js",
     "build:helper": "node build/build-helper.js",
     "build:indices": "node build/build-indices.js",
+    "update:version": "node build/update-version.js",
     "build:mono": "rimraf dist/components && node build/monorepo.js",
     "build:prod": "sh scripts/monorepo.sh",
     "build:directives": "cd packages/directives && yarn clean && yarn build",
@@ -21,6 +22,7 @@
     "build:locale": "cd packages/locale && yarn clean && yarn build",
     "build:theme": "cd packages/theme-chalk && yarn clean && yarn build",
     "build:utils": "cd packages/utils && yarn clean && yarn build",
+    "build:tokens": "cd packages/tokens && yarn clean && yarn build",
     "format": "yarn format:scss",
     "format:scss": "prettier --write **/*.scss",
     "lint": "eslint ./packages --ext .vue,.js,.ts",
