+++ packages/components/popper/index.ts
@@ -14,6 +14,7 @@ export const ElPopper = _Popper
 
 export { default as popperDefaultProps, Effect } from './src/use-popper/defaults'
 export * from './src/renderers'
-export { default as usePopper } from './src/use-popper/index'
+export { default as usePopper } from './src/use-popper'
 export type { Placement, Options } from '@popperjs/core'
+export type { EmitType } from './src/use-popper'
 export type { TriggerType, IPopperOptions, PopperInstance } from './src/use-popper/defaults'
