+++ packages/components/select-v2/src/select-dropdown.vue
@@ -18,7 +18,7 @@ import OptionItem from './option-item.vue'
 
 import { selectV2InjectionKey } from './token'
 
-import type { ItemProps } from '@element-plus/components/virtual-list/src/types'
+import type { ItemProps } from '@element-plus/components/virtual-list'
 import type { OptionItemProps, Option } from './select.types'
 
 export default defineComponent({
