+++ packages/components/upload/src/index.vue
@@ -10,7 +10,7 @@ import {
   onBeforeUnmount,
 } from 'vue'
 import { NOOP } from '@vue/shared'
-import { elFormKey } from '@element-plus/components/form'
+import { elFormKey } from '@element-plus/tokens'
 
 import ajax from './ajax'
 import UploadList from './upload-list.vue'
@@ -18,7 +18,7 @@ import Upload from './upload.vue'
 import useHandlers from './useHandlers'
 
 import type { PropType } from 'vue'
-import type { ElFormContext } from '@element-plus/components/form'
+import type { ElFormContext } from '@element-plus/tokens'
 import type { Nullable } from '@element-plus/utils/types'
 import type {
   ListType,
