+++ packages/element-plus/components.ts
@@ -1,0 +1,131 @@
+import { ElAffix } from '@element-plus/components/affix'
+import { ElAlert } from '@element-plus/components/alert'
+import { ElAutocomplete } from '@element-plus/components/autocomplete'
+import { ElAvatar } from '@element-plus/components/avatar'
+import { ElBacktop } from '@element-plus/components/backtop'
+import { ElBadge } from '@element-plus/components/badge'
+import { ElBreadcrumb } from '@element-plus/components/breadcrumb'
+import { ElButton } from '@element-plus/components/button'
+import { ElCalendar } from '@element-plus/components/calendar'
+import { ElCard } from '@element-plus/components/card'
+import { ElCarousel } from '@element-plus/components/carousel'
+import { ElCascader } from '@element-plus/components/cascader'
+import { ElCascaderPanel } from '@element-plus/components/cascader-panel'
+import { ElCheckTag } from '@element-plus/components/check-tag'
+import { ElCheckbox } from '@element-plus/components/checkbox'
+import { ElCol } from '@element-plus/components/col'
+import { ElCollapse } from '@element-plus/components/collapse'
+import { ElCollapseTransition } from '@element-plus/components/collapse-transition'
+import { ElColorPicker } from '@element-plus/components/color-picker'
+import { ElConfigProvider } from '@element-plus/components/config-provider'
+import { ElContainer } from '@element-plus/components/container'
+import { ElDatePicker } from '@element-plus/components/date-picker'
+import { ElDescriptions } from '@element-plus/components/descriptions'
+import { ElDialog } from '@element-plus/components/dialog'
+import { ElDivider } from '@element-plus/components/divider'
+import { ElDrawer } from '@element-plus/components/drawer'
+import { ElDropdown } from '@element-plus/components/dropdown'
+import { ElEmpty } from '@element-plus/components/empty'
+import { ElForm } from '@element-plus/components/form'
+import { ElIcon } from '@element-plus/components/icon'
+import { ElImage } from '@element-plus/components/image'
+import { ElImageViewer } from '@element-plus/components/image-viewer'
+import { ElInput } from '@element-plus/components/input'
+import { ElInputNumber } from '@element-plus/components/input-number'
+import { ElLink } from '@element-plus/components/link'
+import { ElMenu } from '@element-plus/components/menu'
+import { ElPageHeader } from '@element-plus/components/page-header'
+import { ElPagination } from '@element-plus/components/pagination'
+import { ElPopconfirm } from '@element-plus/components/popconfirm'
+import { ElPopover } from '@element-plus/components/popover'
+import { ElPopper } from '@element-plus/components/popper'
+import { ElProgress } from '@element-plus/components/progress'
+import { ElRadio } from '@element-plus/components/radio'
+import { ElRate } from '@element-plus/components/rate'
+import { ElResult } from '@element-plus/components/result'
+import { ElRow } from '@element-plus/components/row'
+import { ElScrollbar } from '@element-plus/components/scrollbar'
+import { ElSelect } from '@element-plus/components/select'
+import { ElSelectV2 } from '@element-plus/components/select-v2'
+import { ElSkeleton } from '@element-plus/components/skeleton'
+import { ElSlider } from '@element-plus/components/slider'
+import { ElSpace } from '@element-plus/components/space'
+import { ElSteps } from '@element-plus/components/steps'
+import { ElSwitch } from '@element-plus/components/switch'
+import { ElTable } from '@element-plus/components/table'
+import { ElTabs } from '@element-plus/components/tabs'
+import { ElTag } from '@element-plus/components/tag'
+import { ElTimePicker } from '@element-plus/components/time-picker'
+import { ElTimeSelect } from '@element-plus/components/time-select'
+import { ElTimeline } from '@element-plus/components/timeline'
+import { ElTooltip } from '@element-plus/components/tooltip'
+import { ElTransfer } from '@element-plus/components/transfer'
+import { ElTree } from '@element-plus/components/tree'
+import { ElUpload } from '@element-plus/components/upload'
+
+export default [
+  ElAffix,
+  ElAlert,
+  ElAutocomplete,
+  ElAvatar,
+  ElBacktop,
+  ElBadge,
+  ElBreadcrumb,
+  ElButton,
+  ElCalendar,
+  ElCard,
+  ElCarousel,
+  ElCascader,
+  ElCascaderPanel,
+  ElCheckTag,
+  ElCheckbox,
+  ElCol,
+  ElCollapse,
+  ElCollapseTransition,
+  ElColorPicker,
+  ElConfigProvider,
+  ElContainer,
+  ElDatePicker,
+  ElDescriptions,
+  ElDialog,
+  ElDivider,
+  ElDrawer,
+  ElDropdown,
+  ElEmpty,
+  ElForm,
+  ElIcon,
+  ElImage,
+  ElImageViewer,
+  ElInput,
+  ElInputNumber,
+  ElLink,
+  ElMenu,
+  ElPageHeader,
+  ElPagination,
+  ElPopconfirm,
+  ElPopover,
+  ElPopper,
+  ElProgress,
+  ElRadio,
+  ElRate,
+  ElResult,
+  ElRow,
+  ElScrollbar,
+  ElSelect,
+  ElSelectV2,
+  ElSkeleton,
+  ElSlider,
+  ElSpace,
+  ElSteps,
+  ElSwitch,
+  ElTable,
+  ElTabs,
+  ElTag,
+  ElTimePicker,
+  ElTimeSelect,
+  ElTimeline,
+  ElTooltip,
+  ElTransfer,
+  ElTree,
+  ElUpload,
+]
