+++ packages/element-plus/defaults.ts
@@ -1,0 +1,8 @@
+import makeInstaller from './make-installer'
+import Components from './components'
+import Plugins from './plugins'
+
+export default makeInstaller([
+  ...Components,
+  ...Plugins,
+])
