+++ packages/element-plus/package.json
@@ -1,6 +1,6 @@
 {
   "name": "element-plus",
-  "version": "0.0.1",
+  "version": "0.0.4",
   "description": "> TODO: description",
   "author": "Herrington Darkholme <2883231+HerringtonDarkholme@users.noreply.github.com>",
   "homepage": "https://github.com/element-plus/element-plus#readme",
@@ -14,13 +14,15 @@
     "url": "https://github.com/element-plus/element-plus/issues"
   },
   "dependencies": {
-    "@element-plus/components": "^0.0.1",
-    "@element-plus/directives": "^0.0.1",
-    "@element-plus/theme-chalk": "^0.0.1",
-    "@element-plus/hooks": "^0.0.1",
-    "@element-plus/utils": "^0.0.1"
+    "@element-plus/components": "^0.0.4",
+    "@element-plus/directives": "^0.0.4",
+    "@element-plus/theme-chalk": "^0.0.4",
+    "@element-plus/hooks": "^0.0.4",
+    "@element-plus/utils": "^0.0.4",
+    "@element-plus/tokens": "^0.0.4"
   },
   "devDependencies": {
     "@element-plus/test-utils": "^0.0.1"
-  }
+  },
+  "gitHead": "cdeb5a307058d0a2ceccc2848329a5c36e7607bb"
 }
