+++ packages/element-plus/plugins.ts
@@ -1,0 +1,15 @@
+import { ElInfiniteScroll } from '@element-plus/components/infinite-scroll'
+import { ElLoading } from '@element-plus/components/loading'
+import { ElMessage } from '@element-plus/components/message'
+import { ElMessageBox } from '@element-plus/components/message-box'
+import { ElNotification } from '@element-plus/components/notification'
+import { ElPopoverDirective } from '@element-plus/components/popover'
+
+export default [
+  ElInfiniteScroll,
+  ElLoading,
+  ElMessage,
+  ElMessageBox,
+  ElNotification,
+  ElPopoverDirective,
+]
