+++ packages/hooks/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@element-plus/hooks",
-  "description": "",
-  "version": "0.0.1",
+  "description": "Element Plus composables",
+  "version": "0.0.4",
   "main": "index.js",
   "module": "index.js",
   "unpkg": "index.js",
@@ -16,8 +16,9 @@
     "build": "gulp build"
   },
   "dependencies": {
-    "@element-plus/directives": "^0.0.1",
-    "@element-plus/utils": "^0.0.1",
+    "@element-plus/directives": "^0.0.4",
+    "@element-plus/utils": "^0.0.4",
+    "@element-plus/locale": "^0.0.4",
     "@popperjs/core": "^2.4.4"
   },
   "devDependencies": {
@@ -30,5 +31,6 @@
     "sass": "^1.37.0",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.5"
-  }
+  },
+  "gitHead": "cdeb5a307058d0a2ceccc2848329a5c36e7607bb"
 }
