+++ packages/locale/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@element-plus/locale",
-  "version": "0.0.1",
+  "description": "Element Plus locale",
+  "version": "0.0.4",
   "main": "index.js",
   "module": "index.js",
   "unpkg": "index.js",
@@ -27,5 +28,6 @@
     "sass": "^1.37.0",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.5"
-  }
+  },
+  "gitHead": "cdeb5a307058d0a2ceccc2848329a5c36e7607bb"
 }
