+++ packages/test-utils/package.json
@@ -1,12 +1,13 @@
 {
   "name": "@element-plus/test-utils",
   "private": true,
-  "version": "0.0.1",
+  "version": "0.0.4",
   "main": "dist/index.js",
   "license": "MIT",
   "dependencies": {
     "vue": "3.1.1",
     "@vue/test-utils": "^2.0.0-beta.3",
     "lodash": "^4.17.20"
-  }
+  },
+  "gitHead": "cdeb5a307058d0a2ceccc2848329a5c36e7607bb"
 }
