+++ packages/tokens/tsconfig.json
@@ -3,7 +3,9 @@
     "esModuleInterop": true,
     "rootDir": ".",
     "declaration": true,
-    "target": "ES6",
+    "lib": ["ES2020", "DOM"],
+    "skipLibCheck": true,
+    "target": "es6",
     "moduleResolution": "node"
-  }
+  },
 }
